Add new classes for further abstraction. In progress.
