+++ src/main/java/yapbot/util/Parser.java
@@ -1,0 +1,4 @@
+package yapbot.util;
+
+public class Parser {
+}
